refactor(userBook): UserBook 엔티티·리포지토리를 model/userBook 으로 분리
