+++ src/main/java/com/readum/domain/aiChat/service/BookChatSessionSearchService.java
@@ -12,8 +12,8 @@ import com.readum.model.book.entity.Book;
 import com.readum.model.book.repository.BookRepository;
 import com.readum.model.summary.entity.Summary;
 import com.readum.model.summary.repository.SummaryRepository;
-import com.readum.model.user.entity.UserBook;
-import com.readum.model.user.repository.UserBookRepository;
+import com.readum.model.userBook.entity.UserBook;
+import com.readum.model.userBook.repository.UserBookRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
