+++ src/main/java/com/readum/domain/user/userbook/service/UserBookCreateService.java
@@ -8,8 +8,8 @@ import com.readum.domain.user.userbook.dto.UserBookCreateResult;
 import com.readum.domain.user.userbook.exception.UserBookErrorCode;
 import com.readum.model.book.entity.Book;
 import com.readum.model.book.repository.BookRepository;
-import com.readum.model.user.entity.UserBook;
-import com.readum.model.user.repository.UserBookRepository;
+import com.readum.model.userBook.entity.UserBook;
+import com.readum.model.userBook.repository.UserBookRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
