+++ src/main/java/com/readum/domain/user/userbook/service/UserBookSearchService.java
@@ -2,7 +2,7 @@ package com.readum.domain.user.userbook.service;
 
 import com.readum.domain.user.userbook.dto.UserBookSearchItemResult;
 import com.readum.domain.user.userbook.dto.UserBookSearchResult;
-import com.readum.model.user.repository.UserBookRepository;
+import com.readum.model.userBook.repository.UserBookRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
