+++ src/main/java/com/readum/model/userBook/entity/UserBook.java
@@ -1,4 +1,4 @@
-package com.readum.model.user.entity;
+package com.readum.model.userBook.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
