+++ src/test/java/com/readum/domain/aiChat/service/AiChatSessionSearchServiceTest.java
@@ -8,8 +8,8 @@ import com.readum.domain.aiChat.exception.AiChatErrorCode;
 import com.readum.domain.exception.NotFoundException;
 import com.readum.model.aiChat.repository.AiChatSessionRepository;
 import com.readum.model.aiChat.repository.projection.AiChatSessionListProjection;
-import com.readum.model.user.entity.UserBook;
-import com.readum.model.user.repository.UserBookRepository;
+import com.readum.model.userBook.entity.UserBook;
+import com.readum.model.userBook.repository.UserBookRepository;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
