+++ src/test/java/com/readum/domain/summary/service/SummarySearchServiceTest.java
@@ -19,9 +19,9 @@ import com.readum.model.summary.entity.SummaryFixture;
 import com.readum.model.summary.repository.SummaryJobRepository;
 import com.readum.model.summary.repository.SummaryRepository;
 import com.readum.model.book.repository.BookRepository;
-import com.readum.model.user.entity.UserBook;
-import com.readum.model.user.entity.UserBookFixture;
-import com.readum.model.user.repository.UserBookRepository;
+import com.readum.model.userBook.entity.UserBook;
+import com.readum.model.userBook.entity.UserBookFixture;
+import com.readum.model.userBook.repository.UserBookRepository;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
