+++ src/test/java/com/readum/domain/user/userbook/service/UserBookCreateServiceTest.java
@@ -8,10 +8,10 @@ import com.readum.domain.user.userbook.dto.UserBookCreateResult;
 import com.readum.domain.user.userbook.exception.UserBookErrorCode;
 import com.readum.model.book.entity.Book;
 import com.readum.model.book.entity.BookFixture;
-import com.readum.model.user.entity.UserBook;
-import com.readum.model.user.entity.UserBookFixture;
+import com.readum.model.userBook.entity.UserBook;
+import com.readum.model.userBook.entity.UserBookFixture;
 import com.readum.model.book.repository.BookRepository;
-import com.readum.model.user.repository.UserBookRepository;
+import com.readum.model.userBook.repository.UserBookRepository;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
