+++ src/test/java/com/readum/model/aiChat/repository/AiChatSessionRepositoryTest.java
@@ -7,8 +7,8 @@ import com.readum.model.summary.entity.Summary;
 import com.readum.model.summary.entity.SummaryJobFixture;
 import com.readum.model.summary.repository.SummaryJobRepository;
 import com.readum.model.summary.repository.SummaryRepository;
-import com.readum.model.user.entity.UserBook;
-import com.readum.model.user.repository.UserBookRepository;
+import com.readum.model.userBook.entity.UserBook;
+import com.readum.model.userBook.repository.UserBookRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
