+++ src/test/java/com/readum/model/summary/repository/SummaryRepositoryTest.java
@@ -7,8 +7,8 @@ import com.readum.model.book.repository.BookRepository;
 import com.readum.model.summary.entity.Summary;
 import com.readum.model.summary.entity.SummaryFixture;
 import com.readum.model.summary.repository.projection.SummaryHistoryProjection;
-import com.readum.model.user.entity.UserBook;
-import com.readum.model.user.repository.UserBookRepository;
+import com.readum.model.userBook.entity.UserBook;
+import com.readum.model.userBook.repository.UserBookRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
