+++ src/test/java/com/readum/presentation/controller/aiChat/AiChatStreamGuardrailTest.java
@@ -12,9 +12,9 @@ import com.readum.model.book.entity.Book;
 import com.readum.model.book.entity.BookFixture;
 import com.readum.model.book.repository.BookRepository;
 import com.readum.model.user.entity.User;
-import com.readum.model.user.entity.UserBook;
-import com.readum.model.user.entity.UserBookFixture;
-import com.readum.model.user.repository.UserBookRepository;
+import com.readum.model.userBook.entity.UserBook;
+import com.readum.model.userBook.entity.UserBookFixture;
+import com.readum.model.userBook.repository.UserBookRepository;
 import com.readum.model.user.repository.UserRepository;
 import com.readum.presentation.controller.aiChat.dto.SendMessageRequest;
 import org.junit.jupiter.api.BeforeEach;
